[feat/#107] 옷 추천 기능 구현(현재온도 기반)
- 현재 기온 기준으로 카테고리마다 옷의 범위 중에 현재 기온이 포함되는 온도를 가진 옷들을 추천.

+++ src/main/java/com/clokey/server/domain/recommendation/api/RecommendationController.java
@@ -30,7 +30,7 @@ public class RecommendationController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "HOME_200", description = "성공적으로 조회되었습니다."),
     })
     public BaseResponse<RecommendationResponseDTO.DailyClothesResult> recommendClothes(@Parameter(name = "user",hidden = true) @AuthUser Member member,
-                                                                                   @RequestParam @Valid Float nowTemp) {
+                                                                                   @RequestParam @Valid Double nowTemp) {
         RecommendationResponseDTO.DailyClothesResult response = recommendationService.getRecommendClothes(member.getId(), nowTemp);
         return BaseResponse.onSuccess(SuccessStatus.HOME_SUCCESS, response);
     }
